Alex

+++ srcs/draw/rays.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   rays.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: saperrie <saperrie@student.42.fr>          +#+  +:+       +#+        */
+/*   By: aljulien <aljulien@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 19:55:31 by saperrie          #+#    #+#             */
-/*   Updated: 2024/10/15 22:48:19 by saperrie         ###   ########.fr       */
+/*   Updated: 2024/10/22 11:06:29 by aljulien         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,7 @@
 
 int	raycast(char **map, t_player *player, t_pplane *pplane)
 {
-	float	player_position;
-	float	ray;
+	double	player_position;
+	double	ray;
 	int		columns;
-
-
-
 }
